Adjust upload file configuration and calibration
